+++ packages/alchemy/src/GitHub/index.ts
@@ -1,3 +1,6 @@
+export * as Auth from "./AuthProvider.ts";
 export * from "./Comment.ts";
+export { GitHubCredentials, fromEnv, fromToken } from "./Credentials.ts";
+export * from "./Providers.ts";
 export * from "./Secret.ts";
 export * from "./Variable.ts";